+++ personal-hub-web/src/components/ui/index.ts
@@ -1,7 +1,6 @@
 export { default as UiDialog } from './UiDialog.vue'
 export { default as UiInput } from './UiInput.vue'
 export { default as UiButton } from './UiButton.vue'
-export { default as UiCard } from './UiCard.vue'
 export { default as DialogFooterActions } from './DialogFooterActions.vue'
 export { default as DialogTitleField } from './DialogTitleField.vue'
 export { default as DialogPropCard } from './DialogPropCard.vue'